+++ quilt-rs/src/io/remote/object/put_and_request_checksum.rs
@@ -10,7 +10,7 @@ use crate::error::ChecksumError;
 use crate::error::S3ErrorKind;
 use crate::io::remote::HostChecksums;
 use crate::io::remote::HostConfig;
-use crate::io::remote::s3::s3_error_or_login;
+use crate::io::remote::s3::s3_error_or_session_loss;
 use crate::object_hash::Crc64Hash;
 use crate::object_hash::ObjectHash;
 use crate::object_hash::Sha256ChunkedHash;
@@ -54,7 +54,7 @@ pub async fn put_and_request_checksum(
         .send()
         .await
         .map_err(|err| {
-            s3_error_or_login(err, host_config.host.as_ref(), S3ErrorKind::UploadFile)
+            s3_error_or_session_loss(err, host_config.host.as_ref(), S3ErrorKind::UploadFile)
         })?;
     let checksum = match host_config.checksums {
         HostChecksums::Sha256Chunked => extract_sha256_checksum(&response)?,